scripts/requirements.txt
mysql-connector-python>=8.0.0
requests>=2.25.0
